+++ src/main/java/model/order/DeployOrder.java
@@ -1,9 +1,10 @@
-package model;
+package model.order;
 
 public class DeployOrder extends Order {
     public DeployOrder() {
         super();
         setType("deploy");
     }
+
 
 }
