+++ src/test/java/model/OrderTest.java
@@ -1,10 +1,9 @@
 package model;
 
+import model.order.Order;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
-
-import static org.junit.Assert.*;
 
 public class OrderTest extends Order {
